requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.0.1
django-configurations==2.4.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.11.1
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-dataclass==8.5.14
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
py==1.11.0
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.0.275
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typeguard==4.0.0
typing-inspect==0.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.6
zipp==3.15.0
